--- packages/components/src/ui/Nav/enum.ts
@@ -1,4 +1,0 @@
-export enum Dodam_ETheme {
-  LIGHT = "LIGHT",
-  DARK = "DARK",
-}
