+++ packages/hooks/src/index.ts
@@ -2,3 +2,4 @@ export * from "./useQueryInvalidates";
 export * from "./useDetectThemeMode";
 export * from "./useDatePicker";
 export * from "./useNavBar";
+export * from "./useTheme";
